persistent data: small cleanup in dm_block_manager structure
Eliminate sectors_per_block member (easily determined from block_size).
Move dm_io_client member so that 4byte hole moves to the end of 'struct
block_manager'.
Add missing #ifdef DEBUG in lock_internal().
Added FIXME to superblock_all_zeroes().

+++ drivers/md/persistent-data/dm-block-manager.h
@@ -22,7 +22,7 @@ dm_block_manager_create(struct block_device *bdev,
 			unsigned cache_size);
 void dm_block_manager_destroy(struct dm_block_manager *bm);
 
-size_t dm_bm_block_size(struct dm_block_manager *bm);
+unsigned dm_bm_block_size(struct dm_block_manager *bm);
 dm_block_t dm_bm_nr_blocks(struct dm_block_manager *bm);
 
 /*
